tests: Amend release stale port binding test for RBAC
The current version of the test attempts to simulate chassis
registration prior to starting `ovn-controller`, however it does
not set the `hostname` field.

The RBAC role for `ovn-controller` does not allow for a chassis to
change its own name or hostname, which makes sense as this is used
for authentication.

Update the test to set the `hostname` field when simulating chassis
registration so that `ovn-controller` does not attempt to update it
and subsequently make the test fail.

Fixes b6b3823d4 ("ovn-controller: Fix I-P for SB Port_Binding and OVS Interface")