Introduce representor port plugging support
Introduce plugging module that adds and removes ports on the
integration bridge, as directed by Port_Binding options.

Traditionally it has been the CMSs responsibility to create Virtual
Interfaces (VIFs) as part of instance (Container, Pod, Virtual
Machine etc.) life cycle, and subsequently manage plug/unplug
operations on the Open vSwitch integration bridge.

With the advent of NICs connected to multiple distinct CPUs we can
have a topology where the instance runs on one host and Open
vSwitch and OVN runs on a different host, the smartnic CPU.

The act of plugging and unplugging the representor port in Open
vSwitch running on the smartnic host CPU would be the same for
every smartnic variant (thanks to the devlink-port[0][1]
infrastructure) and every CMS (Kubernetes, LXD, OpenStack, etc.).
As such it is natural to extend OVN to provide this common
functionality through its CMS facing API.

The instance will be connected to a SR-IOV Virtual Function or a
RDMA Mediated Device on the host system (the latter not currently
addressed in this implementation). The smartnic driver will
maintain a representor port for each of the host visible devices
on the smartnic CPU side.

It is the CMSs responsibility to maintain a mapping between
instance host and smartnic host, OVN can help by optionally
providing details such as board serial number of the smartnic
system as part of Chassis registration.

The CMS will use it's knowledge of instance host <-> smartnic
host mapping to add appropriate `requested-chassis` along with
the information OVN needs to identify the representor port as
options when creating Logical Switch Ports for instances. These
options will be copied over to the Port_Binding table by
ovn-northd.

OVN will use the devlink-port[0][1] interface to look up which
representor port corresponds to the host visible resource and
maintain presence of this representor port on the integration
bridge.

0: https://www.kernel.org/doc/html/latest/networking/devlink/devlink-port.html
1: https://mail.openvswitch.org/pipermail/ovs-dev/2021-May/382701.html

TODO:

* Implement run-time refresh or incremental updates to
  devlink_ports hash by opening a netlink socket and monitoring
  devlink changes.

* Make use of an index for unbound ports.

* Make use of the existing Incremental Processing support for
  Port-Bindings and call into plugging module from the binding
  module instead of having a separate run loop.

* Implement functionality to convey smartnic identifying
  characteristics as external-ids on Chassis registration. The
  intention is to help CMS map relationships between hosts. Source
  can be board serial number obtained from devlink-info, PCI VPD,
  or a value stored in the local Open vSwitch database maintained
  by the operator.

* Write tests.

* Update documentation.