ovn-sb: Add requested_chassis column to Port_Binding.
To allow for ovn-controller to efficiently monitor all
Port_Binding records destined to it, we add a new
requested_chassis column with weakRef to the Chassis table.

The ovn-controller can monitor this column and only process
records for its chassis UUID before having claimed the port.

northd will fill this column with UUID of Chassis referenced
in Logical_Switch_Port options:requested-chassis by name or
hostname.

In a subsequent update to the controller we will improve the
efficiency of the requested-chassis feature by using the new
column instead of each chassis performing string comparison on
this option for every Port_Binding record processed.

Note that the CMS facing Northbound Logical_Switch_Port:options
API remains the same.