ovn-sb: Add plugged_by column to Port_Binding.
To allow for use of optional plugging support we add a new
plugged_by column with weakRef to the Chassis table.  The
ovn-controller can monitor this column and only process events
for its chassis UUID.

ovn-northd will fill this column with UUID of Chassis referenced
in Logical_Switch_Port options:requested-chassis when
options:plug-type is defined.