ovn-sb: Add requested_chassis column to Port_Binding.
To allow for ovn-controller to efficiently process unbound ports
that may be destined to it, for example for use in the optional
plugging support, we add a new requested_chassis column with
weakRef to the Chassis table.  The ovn-controller can monitor
this column and only process events for its chassis UUID even
before a local binding appears.

northd will fill this column with UUID of Chassis referenced
in Logical_Switch_Port options:requested-chassis by name or
hostname.

Deprecate the OVN_Southbound:Port_Binding:options
"requested-chassis" key.  In a subsequent update to the
controller we will improve the efficiency of the requested-chassis
feature by using the new column instead of each chassis performing
option processing and string comparison.

Note that the CMS facing Northbound Logical_Switch_Port:options
API remains the same.