binding: Consider plugging of ports on CMS request
When OVN is linked with an appropriate plugging implementation,
CMS can request OVN to plug individual lports into the local
Open vSwitch instance.

The port and instance record will be maintained during the lifetime
of the lport and it will be removed on release of lport.

TODO: The functions considering plugging or unplugging of
      interfaces are currently oblivious to the fact that the
      binding module has data structures to keep track of these.
      The functions should be good citizens and learn to update
      those appropriately as they take action.