ovn-controller: Prepare plugging infrastructure.
Add port by interfaces index - To be able to effectively remove
ports previously plugged by us we need to look up ports by
interface records.

Add `enable-dummy-plug` option - To enable testing of the plugging
infrastructure without building OVN with an external plugging
library we include a dummy implementation which can be enabled
using this command line option.

Also adds a ENGINE_NODE_NO_DATA macro which is useful for when
lookup data is located elsewhere and not passed around as part of
the incremental processsing engine.