tests: Stop using non-existent requested-chassis
A couple of tests refers to non-existent chassis to validate that
the ovn-controller would release ports that otherwise was destined
for it when requested-chassis points somewhere else.

With upcoming changes to have the ovn-controller use the new
Port_Binding:requested_chassis column maintained by northd instead
of processing the options itself, referring to a non-existant
chassis will no longer work as the tests expect.

With this patch we instead add a throw-away chassis to the
simulation and point the requested-chassis option at that chassis.