plug_providers: Introduce representor plugin.
Add the first in-tree plug provider plugin and its dependencies.
The representor plugin can be used with multiple NIC vendors
supporting Open vSwitch hardware offload and the devlink-port
infrastructure[0].

It is particularly useful for use with NICs connected to multiple
distinct CPUs where the instance runs on one host and Open
vSwitch and OVN runs on a different host, the smartnic CPU.

Extend the build system with macros from the OVS build system to
allow checking for dependencies of the plugin as well as providing
kernel header files that may not be available at build time.

The plugin will only be built when enabled and when building on
a Linux system.

0: https://www.kernel.org/doc/html/latest/networking/devlink/devlink-port.html