binding: Make local_binding data structure public
The binding module maintains a shash of lports bound to the local
chassis.  Other modules have interest in the same data for lookup,
and it would be wasteful to reimplement the same data structure
elsewhere.

The incremental processing engine already makes sharing of the
data between nodes safe and convenient and this change makes it
possible for a different module to consume the data.