controller: Use multicast for IPv6 Prefix Delegation.
The OVN Controller currently uses unicast messages when
communicating with the DHCPv6 server.

This is at odds with RFC 3315 paragraph 1 in each of the sections
18.2.1, 18.2.3, 18.2.6, and 18.2.7.

The client is expected to use multicast, unless the server has
explicitly told it to use unicast using the Server Unicast
Option [0].  I contemplated adding support for this option, but
I found no server implementation that made use of it.

Other supporting documentation is a change to ISC DHCP server
making it reject unicast Request, Renew, Decline and Release
messages [1].  The dnsmasq implementation also enforces the use
of multicast messages [2].

Fix the ovn-controller IPv6 Prefix Delegation implementation by
making it send requests using the all DHCP agents multicast
address [3] as expected.

0: https://datatracker.ietf.org/doc/html/rfc3315#section-22.12
1: https://github.com/isc-projects/dhcp/commit/b31fe1d552
2: https://github.com/imp/dnsmasq/blob/770bce967c/src/rfc3315.c#L347-L357
3: https://www.rfc-editor.org/rfc/rfc2375.html
Fixes: e3a398e9146e ("controller: Add ipv6 prefix delegation state machine")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>