controller: Always populate prefix-length for IPv6 PD.
The prefix we obtain is used to fill the ``ipv6_ra_prefixes``
option for configuration of instances through SLAAC.

As discussed in RFC 7421 the interface identifier is 64 bits long,
and client implementations refrain from performing SLAAC with any
other prefix length.

TODO: I'm pretty sure the origin of this behavior is tied to an
      older RFC, attempt to track it down.

The DHCPv6 server is at liberty to offer whichever prefix length it
wants, so make sure new requests are for a prefix length of 64 so
that it is suitable.

TODO: We may need to make the default prefix-length to request
      configurable?  Setting this to 64 fixes an issue I see in my
      network where the server will delegate a /62 by default,
      which does not work well for SLAAC.

A future improvement might be to adjust the requested prefix length
to the number of downstream LRPs we have and then allocate to the
downstream LRPs from that larger delegated prefix.

TODO: We are free to request whatever prefix-length we want, and
      the server is equally free to ignore our request and give us
      something else.  So we need some generic handling of
      allocations where prefix-length < 64.  Perhaps just split
      the prefix up in /64 chunks and add them all to
      ``ipv6_ra_prefixes`` ?

TODO: tests

QUESTION: From cursory view there appears to be an issue with
          routing inside OVN after successful IPv6 Prefix
          Delegation, it may be an issue with my setup, but
          thought I might as well ask.  Is the prefix delegation
          code supposed to automatically handle internal routing?

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
(cherry picked from commit 0d176355f26a175e2872b3552838aa67218f782a)