tests: Don't assume first bucket will be chosen.
The commit in the fixes tag introduced a system test that configure
load balancers with multiple backends.

Load balancers are implemented with the group action with equally
weighted buckets.

One of the assertions in the added system test are on there being
exactly one backends being utilized, and this assertion assumes
that the first bucket of the group will be used.

As demonstrated by the reported bug, this is not a correct
assumption, and factors such as system architecture may play a
role in which bucket will be selected.

Since the purpose of the assertion is to confirm only one backend
is utilized, we do not need to match on the exact IP address.

Clear the reply src value, allowing us to assert on there being one
data path flow, regardless of which bucket is chosen.

Reported-at: https://launchpad.net/bugs/2104227
Fixes: 15ab046e12cd ("northd: Add ipv6_{src, dst} to selection_fields column in the NB db.")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>