controller: Fix route filtering by port name.
The commit in the fixes tag introduced the an option to filter
learning of routes to certain LRPs of an LR.

However, when an LR has LRPs both with and without the option,
routes are attempted associated with LRPs without the option,
contrary to the documented behavior.

This was hidden in the original dynamic-routing - Gateway Router
test by populating all LRPs with the option, however this does
not represent a real world use case.

Ensure only LRPs with the option are considered when any of them
has the option set.

Fixes: d7d886eca553 ("controller: Support learning routes per iface.")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>